backend/requirements.txt
fastapi
uvicorn
selenium
python-telegram-bot
pandas
codaio
pydantic